backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-decouple==3.8
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
requests==2.31.0
httpx==0.25.2
twilio==8.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
black==23.11.0
flake8==6.1.0
python-dotenv==1.0.0
